backend/requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy
pillow
pytesseract
python-multipart